Add pytest tests to dbt-clickhouse adapter.
The tests require a running clickhouse-server to connect to. By default, it is assumed there's a running ClickHouse server to connect to.

Use RUN_DOCKER_ENV_VAR_NAME=True to raise a docker image that launches a clickhouse-server through env variables (see README file for more details as well as other env variables).

Docker is not a dependency of the tests, it is disabled by default and it is not part of the dev_requirements.txt.

The new tests replace completely the old deprecated test (deleted now).