Atomic view creation in view.sql
This PR changes the view creation to use `create or replace view`, which is atomic.

Previously, dbt-clickhouse would:
1. Create a new "temporary" view
2. Drop the current view (eek!)
3. Rename the new view to the current view

The above 3 steps aren't done in a single transactions, so any queries on the view that occur during the process may get errors that the view doesn't exist.

Ideally, the operations are done in a single, atomic operation. I.e. use `create or replace view`.