Convert `setup.py` to `pyproject.toml`
Move to declarative packaging to provide better tooling
integration and build isolation (by enabling `python-build`).

This will run the build in isolation with the environment and being
explicit about the dependencies that are used to build the sdists and
bdists, making sure that the build is reproducible and isolated.

The check for the version compatibility has been moved to a unit test,
to move any scripting out of the project metadata.